requirements.txt

zipfile
xml.etree.ElementTree
io
pandas
numpy
scipy.signal
plotly.express
json
pathlib
datetime
openpyxl
